Add longer rotation time for outer face
Had some difficulty with ValueAnimator's update listener.  Normally I'd get the current animated value from the valueanimator and update the circle radius via that, but I kept getting 0 for all updates, so the circles were just sitting still.  Instead decided to repurpose the duration offset and use it as an increment offset as well